# docs/openapi/repository-api-v1.json
{
  "components": {
    "schemas": {
...
      "DependencySelectorResponse": {
        "additionalProperties": false,
        "description": "Direct dependency selector returned by resolution.",
        "properties": {
          "markers": {
            "items": {
              "type": "string"
            },
            "title": "Markers",
            "type": "array"
          },
          "optional": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optional"
          },
          "slug": {
            "description": "Stable public slug of the dependency skill.",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
            "title": "Slug",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Exact immutable dependency version.",
            "title": "Version"
          },
          "version_constraint": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Comma-separated semver comparators.",
            "title": "Version Constraint"
          }
        },
        "required": [
          "slug"
        ],
        "title": "DependencySelectorResponse",
        "type": "object"
      },
      "ErrorBody": {
        "description": "Error detail object for the public API error envelope.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured metadata to help clients debug the error.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable summary of the failure.",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "Standardized error envelope shared across HTTP endpoints.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody",
            "description": "Normalized error payload."
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "SkillDependencyResolutionResponse": {
        "description": "Exact direct dependency declarations for one immutable version.",
        "properties": {
          "depends_on": {
            "items": {
              "$ref": "#/components/schemas/DependencySelectorResponse"
            },
            "title": "Depends On",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "version",
          "depends_on"
        ],
        "title": "SkillDependencyResolutionResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Registry-first API for immutable skill publication, candidate discovery,\nexact dependency reads, immutable metadata batch fetch, immutable content batch\nfetch, and governed lifecycle updates.\n\nThe server owns data-local registry operations only. Prompt interpretation,\nreranking, dependency solving, final selection, lock generation, and execution\nplanning remain client-owned behavior outside this API boundary.",
    "title": "aptitude-server Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/resolution/{slug}/{version}": {
      "get": {
        "description": "Return authored direct `depends_on` declarations for one exact immutable version. This route does not recurse, solve version constraints, or return non-dependency relationship families.",
        "operationId": "getDirectDependencies",
        "parameters": [
          {
            "description": "Stable public slug of the skill to resolve.",
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "description": "Stable public slug of the skill to resolve.",
              "pattern": "^[A-Za-z0-9](?:[A-Za-z0-9._-]{0,127})$",
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "Exact immutable semantic version to resolve.",
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "description": "Exact immutable semantic version to resolve.",
              "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$",
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": {
                  "depends_on": [
                    {
                      "markers": [
                        "linux",
                        "gpu"
                      ],
                      "optional": true,
                      "slug": "python.base",
                      "version_constraint": ">=1.0.0,<2.0.0"
                    }
                  ],
                  "slug": "python.lint",
                  "version": "1.2.3"
                },
                "schema": {
                  "$ref": "#/components/schemas/SkillDependencyResolutionResponse"
                }
              }
            },
            "description": "Direct dependency declarations returned successfully."
          },
          "404": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "SKILL_VERSION_NOT_FOUND",
                    "details": {
                      "slug": "python.lint",
                      "version": "9.9.9"
                    },
                    "message": "Skill version not found: python.lint@9.9.9"
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested immutable `slug@version` does not exist."
          },
          "422": {
            "content": {
              "application/json": {
                "example": {
                  "error": {
                    "code": "INVALID_REQUEST",
                    "details": {
                      "errors": [
                        {
                          "input": "latest",
                          "loc": [
                            "path",
                            "version"
                          ],
                          "msg": "String should match pattern",
                          "type": "string_pattern_mismatch"
                        }
                      ]
                    },
                    "message": "Request validation failed."
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The path parameters are invalid."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read direct immutable dependencies",
        "tags": [
          "resolution"
        ]
      }
    },
...
  }
}